@@ -60,53 +60,54 @@
         }
     },
     "devDependencies": {
-        "@babel/core": "^7.11.1",
-        "@babel/preset-env": "^7.11.0",
+        "@babel/core": "^7.11.6",
+        "@babel/preset-env": "^7.11.5",
         "@babel/preset-typescript": "^7.10.4",
-        "@commitlint/cli": "^9.1.1",
-        "@commitlint/config-conventional": "^9.1.1",
-        "@types/jest": "^26.0.9",
-        "@types/node": "^14.0.27",
-        "@typescript-eslint/eslint-plugin": "^2.33.0",
-        "@typescript-eslint/parser": "^2.33.0",
-        "@wdio/cli": "^6.4.0",
-        "@wdio/local-runner": "^6.4.0",
-        "@wdio/mocha-framework": "^6.4.0",
-        "@wdio/spec-reporter": "^6.4.0",
-        "@wdio/sync": "^6.4.0",
+        "@commitlint/cli": "11.0.0",
+        "@commitlint/config-conventional": "11.0.0",
+        "@types/jest": "^26.0.14",
+        "@types/node": "^14.11.2",
+        "@typescript-eslint/eslint-plugin": "^4.2.0",
+        "@typescript-eslint/parser": "^4.2.0",
+        "@wdio/cli": "^6.5.2",
+        "@wdio/local-runner": "^6.5.2",
+        "@wdio/mocha-framework": "^6.5.0",
+        "@wdio/spec-reporter": "^6.4.7",
+        "@wdio/sync": "^6.5.0",
         "babel-jest": "^26.3.0",
-        "chromedriver": "^84.0.1",
-        "commitizen": "^4.1.2",
+        "chromedriver": "^85.0.1",
+        "codecov": "^3.7.2",
+        "commitizen": "^4.2.1",
         "conventional-changelog-cli": "^2.1.0",
         "cspell": "^4.1.0",
         "depcheck": "^1.2.0",
         "doctoc": "^1.4.0",
-        "eslint": "^7.6.0",
+        "eslint": "^7.9.0",
         "eslint-config-prettier": "^6.11.0",
-        "eslint-import-resolver-typescript": "^2.2.0",
+        "eslint-import-resolver-typescript": "^2.3.0",
         "eslint-plugin-eslint-comments": "^3.1.2",
         "eslint-plugin-import": "^2.22.0",
-        "eslint-plugin-jest": "^23.20.0",
+        "eslint-plugin-jest": "^24.0.2",
         "eslint-plugin-license-header": "^0.2.0",
-        "eslint-plugin-markdown": "^2.0.0-alpha.0",
+        "eslint-plugin-markdown": "^2.0.0-rc.0",
         "eslint-plugin-notice": "^0.9.10",
         "eslint-plugin-prettier": "^3.1.4",
         "eslint-plugin-tsdoc": "^0.2.6",
         "eslint-watch": "^7.0.0",
-        "graphviz-cli": "^1.0.0",
-        "husky": "^4.2.5",
-        "jest": "^26.4.0",
+        "husky": "^4.3.0",
+        "jest": "^26.4.2",
         "lerna": "^3.21.0",
-        "lint-staged": "^10.2.2",
+        "lint-staged": "^10.4.0",
         "lockfile-lint": "^4.3.7",
-        "markdown-link-check": "^3.8.1",
+        "markdown-link-check": "^3.8.3",
         "package-dependency-graph": "^1.9.1",
-        "prettier": "^2.0.5",
+        "prettier": "^2.1.2",
         "rimraf": "^3.0.2",
-        "ts-node": "^8.10.2",
-        "typescript": "^3.9.7",
-        "wdio-chromedriver-service": "^6.0.3",
-        "webdriverio": "^6.4.0"
+        "ts-node": "^9.0.0",
+        "typescript": "^4.0.3",
+        "vertioner": "^1.0.6",
+        "wdio-chromedriver-service": "^6.0.4",
+        "webdriverio": "^6.5.2"
     },
     "engines": {
         "node": "^10 || ^12"
