@@ -9,10 +9,12 @@ module.exports = {
     root: true,
     parser: '@typescript-eslint/parser',
     parserOptions: {
-        // Specify a tsconfig specifically for eslint with simple glob patterns without project references.
-        // Without this checking individual files with lint-staged fails.
+        // Specify a tsconfig specifically for eslint with simple glob patterns without project references
+        // as @typescript-eslint does not support project references:
+        // https://github.com/typescript-eslint/typescript-eslint/issues/2094
+        // Without this checking individual files with lint-staged fails:
         // https://github.com/typescript-eslint/typescript-eslint/issues/967#issuecomment-531817014
-        // TODO: Debug why extending tsconfig.common.json in the following file doesn't work
+        tsconfigRootDir: __dirname,
         project: './tsconfig.eslint.json',
     },
     plugins: [
@@ -36,6 +38,7 @@ module.exports = {
         'plugin:import/errors',
         'plugin:import/warnings',
         'plugin:eslint-comments/recommended',
+        'plugin:markdown/recommended',
     ],
     settings: {
         'import/resolver': {
