@@ -7,7 +7,7 @@
 
 import fs from 'fs';
 import path from 'path';
-import { nameSpace } from '../src';
+import { namespace } from '../src';
 import { axeVersion } from '@sa11y/common';
 
 type objectWithVersion = {
@@ -20,24 +20,39 @@ type objectWithVersion = {
  */
 function isLoaded(objName: string): Promise<string | boolean> {
     return browser.execute((objName) => {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-explicit-any
         const obj: objectWithVersion = (window as { [key: string]: any })[objName];
         return typeof obj === 'object' ? obj.version : false;
     }, objName);
 }
 
-describe('@sa11y/browser-lib', () => {
-    it('should inject minified js', () => {
-        const sa11yMinJsPath = path.resolve(__dirname, '../dist/sa11y.min.js');
-        const sa11yMinJs = fs.readFileSync(sa11yMinJsPath).toString();
-        expect(sa11yMinJs.length).toBeGreaterThan(0);
+/**
+ * Test util function to inject given file and verify that sa11y and axe are loaded into browser
+ */
+function verifySa11yLoaded(filePath: string): void {
+    const sa11yMinJsPath = path.resolve(__dirname, filePath);
+    const sa11yMinJs = fs.readFileSync(sa11yMinJsPath).toString();
+    expect(sa11yMinJs.length).toBeGreaterThan(0);
 
-        // Before injecting sa11y min js neither sa11y nor axe should not be defined
-        expect(isLoaded(nameSpace)).toBe(false);
-        expect(isLoaded('axe')).toBe(false);
-        browser.execute(sa11yMinJs);
-        // After injecting sa11y and axe should be defined
-        // TODO (refactor): Get sa11y version dynamically (from package.json)
-        expect(isLoaded(nameSpace)).toEqual('0.1.0-alpha');
-        expect(isLoaded('axe')).toEqual(axeVersion);
+    // Before injecting sa11y min js neither sa11y nor axe should not be defined
+    void browser.reloadSession();
+    expect(isLoaded(namespace)).toBe(false);
+    expect(isLoaded('axe')).toBe(false);
+    void browser.execute(sa11yMinJs);
+    // After injecting sa11y and axe should be defined
+    // TODO (refactor): Get sa11y version dynamically (from package.json)
+    expect(isLoaded(namespace)).toEqual('0.1.1-alpha.0');
+    expect(isLoaded('axe')).toEqual(axeVersion);
+}
+
+describe('@sa11y/browser-lib', () => {
+    // eslint-disable-next-line jest/expect-expect
+    it('should inject minified js', () => {
+        verifySa11yLoaded('../dist/sa11y.min.js');
+    });
+
+    // eslint-disable-next-line jest/expect-expect
+    it('should inject un-minified js', () => {
+        verifySa11yLoaded('../dist/sa11y.js');
     });
 });
