@@ -9,4 +9,4 @@ export { base, recommended, full } from '@sa11y/preset-rules';
 export { exceptionListFilter } from '@sa11y/format';
 export { assertAccessible } from '@sa11y/assert';
 
-export const nameSpace = 'sa11y';
+export const namespace = 'sa11y';
