@@ -20,19 +20,14 @@ describe('jest setup', () => {
     });
 
     it('should throw error when global expect is undefined', () => {
-        /* eslint-disable @typescript-eslint/ban-ts-ignore */
-        // @ts-ignore error TS2339: Property 'expect' does not exist on type 'Global'.
-        const globalExpect = global.expect;
+        const globalExpect = global.expect as jest.Expect;
         expect(globalExpect).toBeDefined();
         expect(registerSa11yMatcher).not.toThrow();
         try {
-            // @ts-ignore error TS2339: Property 'expect' does not exist on type 'Global'.
             global.expect = undefined;
             globalExpect(registerSa11yMatcher).toThrowErrorMatchingSnapshot();
         } finally {
-            // @ts-ignore error TS2339: Property 'expect' does not exist on type 'Global'.
             global.expect = globalExpect;
         }
-        /* eslint-enable @typescript-eslint/ban-ts-ignore */
     });
 });
