@@ -9,7 +9,10 @@ import { BrowserObject, remote } from 'webdriverio';
 import { assertAccessibleSync } from '@sa11y/wdio';
 import { htmlFileWithA11yIssues, htmlFileWithNoA11yIssues } from '@sa11y/test-utils';
 
+// TODO (chore): Raise issue with WebdriverIO - 'sync' missing 'default' in ts def
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-var-requires
 const sync = require('@wdio/sync').default;
+
 const defaultJestTimeout = 5000; // Is there a way to get this dynamically ?
 
 let browser: BrowserObject;
@@ -42,16 +45,18 @@ afterAll(async () => {
 describe('integration test @sa11y/wdio in sync mode', () => {
     it('should throw error for html with a11y issues', () => {
         expect(browser).toBeTruthy();
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call
         return sync(() => {
-            browser.url(htmlFileWithA11yIssues);
+            void browser.url(htmlFileWithA11yIssues);
             expect(() => assertAccessibleSync(browser)).toThrowErrorMatchingSnapshot();
         });
     });
 
     it('should not throw error for html with no a11y issues', () => {
         expect(browser).toBeTruthy();
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call
         return sync(() => {
-            browser.url(htmlFileWithNoA11yIssues);
+            void browser.url(htmlFileWithNoA11yIssues);
             expect(() => assertAccessibleSync(browser)).not.toThrow();
         });
     });
