@@ -11,10 +11,6 @@ import { A11yConfig, recommended } from '@sa11y/preset-rules';
 import { A11yError } from '@sa11y/format';
 import { getViolations } from '@sa11y/common';
 
-// TODO (refactor): Find a way to declare version into axe namespace
-// eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-// @ts-ignore
-// eslint-disable-next-line import/namespace
 export const axeVersion: string | undefined = axe.version;
 
 /**
@@ -22,10 +18,6 @@ export const axeVersion: string | undefined = axe.version;
  */
 export async function getAxeVersion(driver: BrowserObject): Promise<typeof axeVersion> {
     return driver.execute(() => {
-        // TODO (refactor): Find a way to declare version into axe namespace
-        // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-        // @ts-ignore
-        // eslint-disable-next-line import/namespace
         return typeof axe === 'object' ? axe.version : undefined;
     });
 }
@@ -50,12 +42,12 @@ export async function runAxe(driver: BrowserObject, rules: A11yConfig = recommen
     await loadAxe(driver);
 
     // run axe inside browser and return violations
-    return await driver.executeAsync((rules, done) => {
+    return (await driver.executeAsync((rules, done: CallableFunction) => {
         axe.run(document, rules, (err: Error, results: axe.AxeResults) => {
             if (err) throw err;
             done(results.violations);
         });
-    }, rules);
+    }, rules)) as axe.Result[];
 }
 
 /**
@@ -84,7 +76,7 @@ export async function assertAccessible(
  */
 export function assertAccessibleSync(driver: BrowserObject = browser, rules: A11yConfig = recommended): void {
     // Note: https://github.com/webdriverio/webdriverio/tree/master/packages/wdio-sync#switching-between-sync-and-async
-    driver.call(async () => {
+    void driver.call(async () => {
         await assertAccessible(driver, rules);
     });
 }
